feat: form error handling
displays the error message associated to the error on form submittal. used the Message component of semantic-ui and the component state to help display the message. needed to pass the error property to the Form component and also needed it to be a boolean so used the double bang 'git add .' operator

closes #77